Fix #7: Missing config2.php causes fatal error under PHP 8.0
We have to ensure that config2.php is readable and that `json_decode()`
returned an array; otherwise the following code exhibits undefined
behavior at best, or `TypeError`s under PHP 8.0, so we bail out.

We are aware that there is a race condition regarding `is_readable()`
and `file_get_contents()`, and that actually reading that file might
fail for other reasons, but we detest the `@` operator, and prefer to
have spurious warnings on rare occasions.